Fix issue with nulls in api reply
Turns out an endpoint we invoke 'scatters' nulls in its reply, we need
to strip them, because we need the data beyond those nulls.

If nothing is done [NSString initWithData:encoding] ignores aything
after the first null it finds.

So, we convert nulls to spaces and take advantage of the fact that
[NSString initWithData:encoding] works even with non-null terminated
strings.

We incurr in two penalties:
- memory allocation [data length]
- o(n) traversal to copy data to output array

Ref #7