Fix issues with nulls in reply (android)
Turns out an endpoint we invoke 'scatters' nulls in its reply, we need to strip them, because we need the data beyond those nulls.

If nothing is done new String(....) ignores aything after the first null it finds.

So, we convert nulls to spaces and take advantage of the fact that new String(...) works even with non-null terminated strings.

We incurr in an o(n) traversal to replace nulls with spaces

Ref #8